refactor: PR-B3-2548-W1-PR2 — mode folded into list_instances, force_release_worktree merged into release_worktree(force:true)
refs #2548 (P1 Wave 1, PR-2). MCP tools 29 -> 27.

- `mode` (0 calls/20d MCP; read-only `get` action): removed as a
  standalone tool. Its read side folds into `list_instances`'s response
  as a new top-level `operator_mode: {mode, delegate_to, delegate_scope}`
  field — agents already poll list_instances to observe fleet state, so
  operator availability belongs alongside it. Setting the mode stays
  CLI-only (`agend-terminal mode <active|away|sleep>`), unaffected.
  Scope clarification (issue said "fold into list_instances/health"):
  the `health` MCP tool only has report/clear write actions, no read
  branch to fold into — read either "health" as a modifier describing
  list_instances' existing health-ish fields (chosen) or as license to
  add a new get action to health (rejected: new action = new surface,
  contradicts the tool-count-reduction goal). No new consumer for a
  health-report-enriched mode either (considered, deferred — avoid
  speculative API surface until a real consumer appears).

- `force_release_worktree` (25 calls/20d vs release_worktree's 573):
  collapsed into `release_worktree(force:true)`. Key semantics carried
  over exactly (RED-first, 14 tests in mcp/handlers/worktree/tests.rs):
  - `force:false` (default) is byte-identical to pre-#2548
    `release_worktree` — binding-driven, `.agend-managed` marker-gated,
    zero behavior change.
  - `force:true` computes the target path directly from `agent`+`branch`
    (now a required arg only in this mode) and skips the marker check —
    intentional, since this path exists specifically for stale-state
    recovery where a directory survives after its binding is already
    gone, and the marker check would block exactly that recovery.
  - AUDIT2-002 ownership check (caller must be the worktree's own agent
    or its team orchestrator) rides along on the force path — the one
    new authorization rule this PR adds. force:false intentionally still
    has NO ownership check (pre-existing behavior; extending it there is
    a separate authorization-tightening question, explicitly out of
    scope here).
  `force_release/mod.rs` keeps `rebase_clean_self` + the `gc` submodule
  (still shared with `bind_self(rebase_mode=true)`); its own
  `handle_force_release_worktree` wrapper and 12 direct tests are
  removed, folded into `mcp/handlers/worktree.rs`. `gc.rs`'s own tests
  (metadata-pruning behavior) now drive the merged handler with
  `force: true` instead of the retired standalone function.

`worktree.rs` grew past the 750-LOC handler ceiling from the new tests;
split its `#[cfg(test)] mod tests` / `mod path_traversal_tests` into
sibling `worktree/tests.rs` / `worktree/path_traversal_tests.rs` (the
established file-size-invariant split pattern) rather than grandfather
a new monolith.

Updated registry/pin tests (registered_handler_names_pin,
full_capability_roles_surface_all_27_byte_identical, tool count 29->27),
operator_gate.rs (classify() arm removed for `mode`; `release_worktree`
already covers the AbsolutelyNever class force:true now routes through,
so `force_release_worktree` just drops from the list — 2 tests updated
to match), mcp_proxy.rs retry-safe list, docs (MCP-TOOLS*/FEATURE-worktree*/
RUNBOOK*/FLEET-DEV-PROTOCOL.md/README*), and build_feature_tracker.py
(retired-in-place, per #2547/#2548 precedent). Full-repo unrestricted rg
sweep for dangling references before pushing.

Verification: full `cargo test` diffed (sorted FAILED test names) against
a `git stash`-then-rerun baseline on the same commit — byte-identical
246 pre-existing sandboxed-CI failures both sides, zero new regressions.
`cargo fmt --check` and `cargo clippy --all-targets --features tray -- -D
warnings` both clean.

Agend-Agent: gapfix-dev
Agend-Task: t-20260702232416250669-56872-23
Agend-Branch: refactor/2548-w1-pr2
Agend-Issued-At: 2026-07-02T23:24:31.468316+00:00
