+++ scripts/user_story_audit/build_feature_tracker.py
@@ -123,7 +123,7 @@ CLI_ROWS = [
     ("connect", "Connect external agent", "operator", "register an externally managed agent with backend metadata", "external sessions can use daemon MCP without being daemon-spawned", "connect requires name and backend, optionally working dir and extra args, and registers an external agent.", "CLI", "src/main.rs: Commands::Connect; src/connect.rs; docs/FEATURE-agent-interaction.md", "Connect a test external shell/codex backend and verify it appears as external.", "P2"),
     ("stop", "Stop daemon", "operator", "stop the running daemon cleanly", "the fleet shuts down without orphaning normal state", "stop routes to daemon shutdown for the active AGEND_HOME.", "CLI", "src/main.rs: Commands::Stop; src/daemon/mod.rs", "Start temp daemon; run stop; verify API is unavailable and process exits.", "P0"),
     ("kill", "Kill agent", "operator", "terminate a selected agent", "a stuck or unwanted backend can be removed", "kill accepts an agent name and invokes daemon lifecycle teardown; managed agents may be subject to restart policy.", "CLI", "src/main.rs: Commands::Kill; src/agent/deleting.rs", "Start two agents; kill one; verify registry and PTY disappear.", "P1"),
-    ("mode", "Set operator mode", "operator", "set active/away/sleep with optional delegate scope", "agents can respect operator availability and authority boundaries", "mode accepts active, away, or sleep plus delegate/scope values and stores operator authority state.", "CLI", "src/main.rs: Commands::Mode; src/operator_mode.rs", "Set each mode in temp AGEND_HOME and verify MCP mode get reflects it.", "P1"),
+    ("mode", "Set operator mode", "operator", "set active/away/sleep with optional delegate scope", "agents can respect operator availability and authority boundaries", "mode accepts active, away, or sleep plus delegate/scope values and stores operator authority state.", "CLI", "src/main.rs: Commands::Mode; src/operator_mode.rs", "Set each mode in temp AGEND_HOME and verify list_instances' operator_mode field reflects it (#2548: the MCP `mode get` tool was retired, folded into list_instances).", "P1"),
     ("admin-cleanup-branches", "Cleanup merged branches", "operator", "preview and delete merged local branches", "old branch clutter can be removed safely", "admin cleanup-branches defaults to dry-run; --yes applies deletion of eligible branches.", "CLI", "src/main.rs: AdminCommands::CleanupBranches; src/branch_sweep.rs", "Create clean/squash/stale branch fixtures and compare dry-run vs --yes.", "P2"),
     ("admin-cleanup-zombies", "Cleanup zombie daemons", "operator", "find and reap stale daemon run directories/processes", "dead daemon state does not block new launches", "admin cleanup-zombies filters by --age and prompts unless --yes; Unix uses TERM then KILL grace path.", "CLI", "src/main.rs: AdminCommands::CleanupZombies; src/admin/cleanup_zombies.rs", "Use fixture run dirs/process mocks if available; avoid killing real daemon in-session.", "P1", "Do not run destructive zombie cleanup against the live session."),
     ("capture-backend", "Capture backend output", "operator/developer", "record backend terminal output for diagnostics", "state bugs can be reproduced from captured bytes", "capture backend records a selected backend for a bounded duration.", "CLI", "src/main.rs: CaptureAction::Backend; src/capture.rs", "Run a short capture against a harmless shell backend and inspect emitted cap/meta files.", "P2"),
@@ -196,11 +196,11 @@ MCP_TOOLS: list[tuple[str, str, str, str, str, str, str]] = [
     ("repo", "Repo/worktree operations", "agent/operator", "checkout/release/cleanup/merge repository worktrees", "work happens in isolated daemon-managed branches", "Supports checkout bind=true, cleanup init commits/merged branches, and PR merge.", "src/mcp/tools.rs; src/mcp/handlers/ci/*; docs/FEATURE-worktree.md"),
     ("bind_self", "Bind caller to worktree", "agent", "bind my instance to a branch worktree", "I can safely do repository work outside initial dispatch", "Rejects protected branches; supports repository_path and rebase_mode.", "src/mcp/tools.rs; src/binding.rs; src/worktree_pool.rs"),
     ("release_worktree", "Release bound worktree", "agent/operator", "soft-release a daemon-managed worktree", "worktree ownership and cleanup state are clear", "Clears binding and marks .agend-managed release metadata; dry_run previews.", "src/mcp/tools.rs; src/worktree_pool.rs"),
-    ("force_release_worktree", "Emergency worktree release", "operator", "force clean stale daemon-managed worktrees", "dangling leases can be recovered safely", "Requires instance/branch and refuses paths outside daemon pool.", "src/mcp/tools.rs; src/mcp/handlers/force_release/*"),
+    ("force_release_worktree", "Emergency worktree release", "operator", "force clean stale daemon-managed worktrees", "dangling leases can be recovered safely", "Retired in #2548 — folded into `release_worktree(force:true)` (same behavior: requires instance/branch, refuses paths outside daemon pool, restricted to worktree owner/team orchestrator per AUDIT2-002); row kept for audit-history continuity.", "src/mcp/tools.rs; src/mcp/handlers/worktree.rs; src/mcp/handlers/force_release/*"),
     ("binding_state", "Binding inspection", "agent/operator", "inspect binding/worktree/CI watch state", "agents can find their authoritative worktree", "Reports binding.json, worktree marker, watches, bind-in-flight, cross-branch holders.", "src/mcp/tools.rs; src/mcp/handlers/binding_state.rs"),
     ("gc_dry_run", "Worktree GC preview", "operator", "preview daemon-managed worktree GC candidates", "cleanup can be audited before deletion", "Retired in #2548 (zero calls in 20 days) — moved to `agend-terminal admin gc-dry-run` CLI; row kept for audit-history continuity.", "src/mcp/tools.rs; src/worktree_pool/gc.rs"),
     ("tokens", "Token/cost summary", "operator", "summarize estimated Claude/Codex token usage", "fleet cost/context consumption can be monitored", "Retired in #2548 (zero calls in 20 days) — moved to `agend-terminal admin tokens` CLI; row kept for audit-history continuity.", "src/mcp/tools.rs; src/token_cost.rs"),
-    ("mode", "Read operator mode", "agent", "read active/away/sleep operator mode", "agents can back off when operator is unavailable", "MCP mode is read-only and returns delegate/scope.", "src/mcp/tools.rs; src/operator_mode.rs"),
+    ("mode", "Read operator mode", "agent", "read active/away/sleep operator mode", "agents can back off when operator is unavailable", "Retired in #2548 — read side folded into `list_instances`'s `operator_mode` field (mode/delegate_to/delegate_scope); setting stays CLI-only. Row kept for audit-history continuity.", "src/mcp/tools.rs; src/mcp/handlers/instance_queries.rs; src/operator_mode.rs"),
 ]
 
 
@@ -340,7 +340,7 @@ TEST_RESULTS: dict[str, tuple[str, str, str, str, str]] = {
     "US-MCP-034": ("Pass", "", "N/A", "Pass", "binding_state returned full structured bind report."),
     "US-MCP-035": ("Pass", "", "N/A", "Pass", "gc_dry_run listed 0 candidates without mutation; tool retired in #2548 — moved to admin gc-dry-run CLI."),
     "US-MCP-036": ("Pass", "", "N/A", "Pass", "tokens summary returned backend cost summary; tool retired in #2548 — moved to admin tokens CLI."),
-    "US-MCP-037": ("Pass", "", "N/A", "Pass", "mode get returned operator mode (read-only)."),
+    "US-MCP-037": ("Pass", "", "N/A", "Pass", "mode get returned operator mode (read-only); tool retired in #2548 — read side folded into list_instances' operator_mode field."),
     # tools/list surface
     "US-SYS-019": ("Pass", "", "N/A", "Pass", "agend-mcp-bridge tools/list returned all 37 tools; tools/call round-tripped."),
     # verify probe-backed system rows
@@ -416,7 +416,7 @@ TEST_RESULTS.update({
     "US-MCP-017": ("Pass", "", "N/A", "Pass", "tui_screenshot correctly errors in daemon-only mode (requires TUI mode) — expected behaviour; tool retired in #2548 (zero calls, no CLI replacement) — the ScreenshotRequest plumbing and src/screenshot.rs were removed."),
     "US-MCP-021": ("Pass", "", "N/A", "Pass", "restart_daemon self-respawned the isolated test daemon (successor confirmed healthy); list showed the agent afterward."),
     "US-MCP-031": ("Pass", "ERR-006 (env-only, not a product defect): bind_self could not be exercised end-to-end because the agend-git shim blocks all git mutations from this agent-rooted shell (ancestry guard #2158/#2234), so a clean source repo cannot be constructed here.", "N/A", "Pass", "MCP handler dispatches and validates (returns structured lease_failed/cross_agent_conflict); binding/worktree_pool unit tests (53+) cover the bind logic."),
-    "US-MCP-033": ("Pass", "", "N/A", "Pass", "force_release_worktree returned a structured binding_outcome (already_released) without mutating non-pool paths."),
+    "US-MCP-033": ("Pass", "", "N/A", "Pass", "force_release_worktree returned a structured binding_outcome (already_released) without mutating non-pool paths; tool retired in #2548 — folded into release_worktree(force:true), same behavior."),
 })
 
 # Third wave: final remaining stories (service lifecycle, feature-gated builds,
