+++ src/mcp/handlers/worktree.rs
@@ -169,15 +169,21 @@ pub(crate) fn handle_bind_self(home: &Path, args: &Value, sender: &Option<Sender
 /// `git worktree remove --force` (or fallback). Idempotent (#1465) — a
 /// second call (no binding) returns `released:true, already_released:true`
 /// (success no-op, no error).
-pub(crate) fn handle_release_worktree(
-    home: &Path,
-    args: &Value,
-    _sender: &Option<Sender>,
-) -> Value {
+///
+/// `force:true` (#2548 PR-2) absorbs the former standalone
+/// `force_release_worktree` tool — see [`handle_release_worktree_force`].
+/// `force:false` (default) is this original binding-driven path,
+/// byte-identical to pre-#2548 behavior.
+pub(crate) fn handle_release_worktree(home: &Path, args: &Value, sender: &Option<Sender>) -> Value {
     let agent = match args["instance"].as_str() {
         Some(a) if !a.is_empty() => a,
         _ => return json!({"error": "missing 'instance'"}),
     };
+
+    if args["force"].as_bool().unwrap_or(false) {
+        return handle_release_worktree_force(home, args, agent, sender);
+    }
+
     crate::validate_name_or_err!(agent);
     let dry_run = args["dry_run"].as_bool().unwrap_or(false);
     // #789: clean empty init commits before removal (best-effort).
@@ -192,355 +198,110 @@ pub(crate) fn handle_release_worktree(
     serde_json::to_value(&outcome).unwrap_or_else(|_| json!({"error": "serialize failed"}))
 }
 
-#[cfg(test)]
-#[allow(clippy::unwrap_used)]
-mod tests {
-    use super::*;
-
-    fn tmp_home(suffix: &str) -> std::path::PathBuf {
-        let h = std::env::temp_dir().join(format!(
-            "agend-p0x-handler-{}-{}",
-            std::process::id(),
-            suffix
-        ));
-        std::fs::create_dir_all(&h).ok();
-        h
+/// `release_worktree(force:true)` (#2548 PR-2) — absorbed from the former
+/// standalone `force_release_worktree` MCP tool. Cleans a stale daemon-
+/// managed worktree directory directly via
+/// `<home>/worktrees/<agent>/<branch>/`, bypassing the `.agend-managed`
+/// marker check the `force:false` path enforces — intentional: this path
+/// exists specifically for stale-state recovery where a directory survives
+/// after its binding is already gone (see `rebase_clean_self`'s docs), and
+/// the marker check would block exactly that recovery. Requires `branch`
+/// (the `force:false` path derives it from the binding; here there may be
+/// no binding left to read).
+///
+/// AUDIT2-002: since this path deletes disk state without the marker safety
+/// net, restrict callers to the target's own agent or its team orchestrator.
+/// An anonymous caller (operator-direct) keeps full authority.
+fn handle_release_worktree_force(
+    home: &Path,
+    args: &Value,
+    agent: &str,
+    sender: &Option<Sender>,
+) -> Value {
+    if let Err(e) = crate::agent::validate_name(agent) {
+        return json!({"error": e, "code": "invalid_agent"});
     }
-
-    #[test]
-    fn handler_rejects_missing_agent() {
-        let home = tmp_home("no-agent");
-        let result = handle_release_worktree(&home, &json!({}), &None);
-        assert_eq!(
-            result["error"].as_str(),
-            Some("missing 'instance'"),
-            "missing instance must surface clear error: {result}"
-        );
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn handler_rejects_invalid_agent_name() {
-        let home = tmp_home("bad-name");
-        // Agent names with `..` are rejected by validate_name.
-        let result = handle_release_worktree(&home, &json!({"instance": "../etc/passwd"}), &None);
-        assert!(
-            result.get("error").is_some(),
-            "invalid agent name must error: {result}"
-        );
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn handler_idempotent_no_binding_returns_success_noop() {
-        // #1465: no binding → idempotent SUCCESS no-op (released:true,
-        // already_released:true, no error; was released:false pre-#1465).
-        let home = tmp_home("idem-no-binding");
-        let result = handle_release_worktree(&home, &json!({"instance": "ghost"}), &None);
-        assert_eq!(result["released"].as_bool(), Some(true), "{result}");
-        assert_eq!(result["already_released"].as_bool(), Some(true), "{result}");
-        assert!(
-            result.get("error").is_none(),
-            "no-op must not error: {result}"
-        );
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    // ── Sprint 54 P1-7: bind_self handler tests ─────────────────────────
-    //
-    // These exercise `handle_bind_self` directly — same path the MCP layer
-    // uses. The helper sets up a real git repo + fleet.yaml entry so
-    // `worktree_pool::lease` can actually create the worktree (matches
-    // dispatch_hook/tests.rs setup_test_repo).
-    //
-    // Regression-proof anchor: replace the body of
-    // `dispatch_auto_bind_lease` with `Ok(())` (skip the actual bind) →
-    // `bind_self_creates_binding_and_worktree` fails because binding.json
-    // never gets written. PR description carries the captured FAIL
-    // signature.
-
-    fn p17_setup_repo(home: &std::path::Path, agent: &str) -> std::path::PathBuf {
-        let repo = crate::paths::workspace_dir(home).join(agent);
-        std::fs::create_dir_all(&repo).ok();
-        std::process::Command::new("git")
-            .args(["init", "-b", "main"])
-            .current_dir(&repo)
-            .env("AGEND_GIT_BYPASS", "1")
-            .output()
-            .ok();
-        std::process::Command::new("git")
-            .args([
-                "-c",
-                "user.name=test",
-                "-c",
-                "user.email=t@t",
-                "commit",
-                "--allow-empty",
-                "-m",
-                "init",
-            ])
-            .current_dir(&repo)
-            .env("AGEND_GIT_BYPASS", "1")
-            .output()
-            .ok();
-        // #781 Phase 3 r1 (Path A): populate `refs/remotes/origin/main`
-        // for strict `ensure_branch_exists`; file:/// URL so
-        // derive_repo returns None.
-        let git = |a: &[&str]| -> Option<std::process::Output> {
-            std::process::Command::new("git")
-                .args(a)
-                .current_dir(&repo)
-                .env("AGEND_GIT_BYPASS", "1")
-                .output()
-                .ok()
-        };
-        git(&["remote", "add", "origin", "file:///dev/null/agend-fix"]);
-        if let Some(o) = git(&["rev-parse", "HEAD"]).filter(|o| o.status.success()) {
-            let sha = String::from_utf8_lossy(&o.stdout).trim().to_string();
-            git(&["update-ref", "refs/remotes/origin/main", &sha]);
+    let branch = match args["branch"].as_str() {
+        Some(b) if !b.is_empty() => b,
+        _ => return json!({"error": "missing 'branch'"}),
+    };
+    if let Some(caller) = sender.as_ref().map(|s| s.as_str()) {
+        if caller != agent && !crate::teams::is_orchestrator_of(home, caller, agent) {
+            return json!({
+                "error": format!(
+                    "permission denied: '{caller}' cannot force-release '{agent}'s worktree \
+                     (only the owner or its team orchestrator may)"
+                ),
+                "code": "not_owner_or_orchestrator"
+            });
         }
-        std::fs::write(
-            crate::fleet::fleet_yaml_path(home),
-            format!(
-                "instances:\n  {agent}:\n    backend: claude\n    working_directory: {}\n",
-                repo.display()
+    }
+    if !crate::agent_ops::validate_branch(branch) {
+        return json!({
+            "error": format!("invalid branch name '{branch}'"),
+            "code": "invalid_branch"
+        });
+    }
+
+    // Safety: ensure the resolved target is within the worktrees pool AND
+    // deeper than the agent-level subdirectory (a `branch == ""` would
+    // otherwise resolve to the agent's own dir; the empty-string check
+    // above already rejects this, but the defense-in-depth guard catches
+    // future validator drift).
+    let worktrees_root = home.join("worktrees");
+    let target = worktrees_root.join(agent).join(branch);
+    let safe = target.starts_with(&worktrees_root)
+        && target != worktrees_root
+        && target != worktrees_root.join(agent);
+    if !safe {
+        return json!({
+            "error": format!(
+                "release_worktree(force:true) refuses to clean path outside the daemon \
+                 worktree pool: {}",
+                target.display()
             ),
-        )
-        .ok();
-        repo
+            "code": "path_outside_pool"
+        });
     }
 
-    fn sender_for(name: &str) -> Option<crate::identity::Sender> {
-        crate::identity::Sender::new(name)
-    }
+    // #826: optional operator-supplied `repository_path` arg. When present,
+    // L2 GC skips enumeration and goes straight to the named repo.
+    let source_repo_hint = args["repository_path"]
+        .as_str()
+        .filter(|s| !s.is_empty())
+        .map(std::path::PathBuf::from);
 
-    #[test]
-    fn bind_self_creates_binding_and_worktree() {
-        // Gate 1: a successful bind_self produces binding.json + worktree
-        // dir + .agend-managed marker. Mirrors the dispatch hook's
-        // happy path because we go through the same helper.
-        //
-        // EMPIRICAL REGRESSION-PROOF ANCHOR: replacing
-        // `dispatch_auto_bind_lease` body with `Ok(())` makes this test
-        // fail with "binding.json must exist after bind_self".
-        let home = std::env::temp_dir().join(format!("agend-p17-self-{}-ok", std::process::id()));
-        std::fs::create_dir_all(&home).ok();
-        p17_setup_repo(&home, "agent-self");
-
-        let resp = handle_bind_self(
-            &home,
-            &json!({"repository": "owner/name", "branch": "feat/p17"}),
-            &sender_for("agent-self"),
-        );
-        assert_eq!(
-            resp["bound"].as_bool(),
-            Some(true),
-            "bind_self must succeed: {resp}"
-        );
-        let worktree_path = resp["worktree_path"]
-            .as_str()
-            .expect("worktree_path in success response");
-        assert!(!worktree_path.is_empty(), "worktree_path must be populated");
-
-        let binding_path = crate::paths::runtime_dir(&home)
-            .join("agent-self")
-            .join("binding.json");
-        assert!(
-            binding_path.exists(),
-            "binding.json must exist after bind_self"
-        );
-        let v: Value =
-            serde_json::from_str(&std::fs::read_to_string(&binding_path).expect("read binding"))
-                .expect("parse binding");
-        assert_eq!(v["branch"].as_str(), Some("feat/p17"));
-        assert_eq!(
-            v["task_id"].as_str(),
-            Some(""),
-            "self-bind without task_id arg must record empty task_id"
-        );
-
-        // Worktree dir + .agend-managed marker per P0-X / P1-7.
-        let wt = std::path::Path::new(worktree_path);
-        assert!(wt.exists(), "worktree dir must exist: {worktree_path}");
-        assert!(
-            wt.join(".agend-managed").exists(),
-            ".agend-managed marker must exist: {worktree_path}"
-        );
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn bind_self_idempotent_same_agent_same_branch() {
-        // Gate 2: a second bind_self call from the same agent on the
-        // same branch is idempotent. The first lease creates the
-        // worktree; the second sees the existing daemon-managed
-        // worktree on the matching branch and succeeds without
-        // mutating state.
-        let home = std::env::temp_dir().join(format!("agend-p17-self-{}-idem", std::process::id()));
-        std::fs::create_dir_all(&home).ok();
-        p17_setup_repo(&home, "agent-idem");
-
-        let args = json!({"repository": "owner/name", "branch": "feat/idem"});
-        let r1 = handle_bind_self(&home, &args, &sender_for("agent-idem"));
-        assert_eq!(r1["bound"].as_bool(), Some(true), "first bind: {r1}");
-        let r2 = handle_bind_self(&home, &args, &sender_for("agent-idem"));
-        assert_eq!(
-            r2["bound"].as_bool(),
-            Some(true),
-            "second bind on same branch must be idempotent: {r2}"
-        );
-        assert_eq!(
-            r1["worktree_path"], r2["worktree_path"],
-            "worktree path must be stable across idempotent calls"
-        );
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn bind_self_rejects_main_branch_with_e4_5() {
-        // Gate 3: protected-branch invariant. Calling bind_self on
-        // 'main' returns the E4.5 rejection from worktree_pool::lease,
-        // mapped to a stable code so agents can branch on it.
-        let home = std::env::temp_dir().join(format!("agend-p17-self-{}-e45", std::process::id()));
-        std::fs::create_dir_all(&home).ok();
-        p17_setup_repo(&home, "agent-e45");
-
-        let resp = handle_bind_self(
-            &home,
-            &json!({"repository": "owner/name", "branch": "main"}),
-            &sender_for("agent-e45"),
-        );
-        assert!(
-            resp.get("error").is_some(),
-            "main branch must error: {resp}"
-        );
-        assert_eq!(
-            resp["code"].as_str(),
-            Some("e4_5_protected_branch"),
-            "error code must surface E4.5 class: {resp}"
-        );
-
-        // No side-effects on rejection.
-        let binding = crate::paths::runtime_dir(&home)
-            .join("agent-e45")
-            .join("binding.json");
-        assert!(
-            !binding.exists(),
-            "rejected bind_self must not write binding.json"
-        );
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn bind_self_rejects_cross_agent_branch_conflict() {
-        // Gate 4: P0-1.5 cross-agent registry — agent A binds, agent B
-        // attempts the same (source_repo, branch) → B is rejected.
-        // #2117 P3b: the lease key is (source_repo, branch). Both agents claim the
-        // SAME repo (via `repository_path`) and branch — the same-repo conflict P3b
-        // preserves. (Cross-repo independence is covered by the dispatch-side
-        // `cross_repo_same_branch_independent_p3b` test.)
-        let home =
-            std::env::temp_dir().join(format!("agend-p17-self-{}-cross", std::process::id()));
-        std::fs::create_dir_all(&home).ok();
-        let shared = p17_setup_repo(&home, "shared-repo");
-        let shared_path = shared.display().to_string();
-
-        let r1 = handle_bind_self(
-            &home,
-            &json!({"repository_path": shared_path, "branch": "feat/cross"}),
-            &sender_for("agent-A"),
-        );
-        assert_eq!(r1["bound"].as_bool(), Some(true), "A binds first: {r1}");
-
-        let r2 = handle_bind_self(
-            &home,
-            &json!({"repository_path": shared_path, "branch": "feat/cross"}),
-            &sender_for("agent-B"),
-        );
-        assert!(
-            r2.get("error").is_some(),
-            "B must be rejected on the same (repo, branch): {r2}"
-        );
-        assert_eq!(
-            r2["code"].as_str(),
-            Some("cross_agent_conflict"),
-            "code must be cross_agent_conflict: {r2}"
-        );
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn bind_self_then_release_worktree_clean_state() {
-        // Gate 5: lifecycle round-trip. bind_self creates state;
-        // release_worktree clears it. binding.json + worktree dir +
-        // .agend-managed marker all gone after release.
-        let home =
-            std::env::temp_dir().join(format!("agend-p17-self-{}-cycle", std::process::id()));
-        std::fs::create_dir_all(&home).ok();
-        p17_setup_repo(&home, "agent-cycle");
-
-        let resp = handle_bind_self(
-            &home,
-            &json!({"repository": "owner/name", "branch": "feat/cycle"}),
-            &sender_for("agent-cycle"),
-        );
-        assert_eq!(resp["bound"].as_bool(), Some(true));
-        let worktree_path = resp["worktree_path"]
-            .as_str()
-            .expect("worktree path")
-            .to_string();
-        let binding = home
-            .join("runtime")
-            .join("agent-cycle")
-            .join("binding.json");
-        assert!(binding.exists());
-
-        let release = handle_release_worktree(&home, &json!({"instance": "agent-cycle"}), &None);
-        assert_eq!(
-            release["released"].as_bool(),
-            Some(true),
-            "release must succeed: {release}"
-        );
-
-        assert!(!binding.exists(), "binding.json must be gone after release");
-        assert!(
-            !std::path::Path::new(&worktree_path).exists(),
-            "worktree dir must be gone after release: {worktree_path}"
-        );
-
-        std::fs::remove_dir_all(&home).ok();
+    match crate::mcp::handlers::force_release::rebase_clean_self(home, agent, branch) {
+        Ok(o) => {
+            // #826 L2 GC: when the binding-clear path short-circuited on
+            // "no binding" (the post-disband state), the
+            // `git worktree remove --force` step inside `release_full`
+            // never ran. Run it now against any source repos that still
+            // hold `.git/worktrees/<meta-dir>/` metadata for our target
+            // worktree path.
+            let gc = crate::mcp::handlers::force_release::prune_git_metadata_for_agent(
+                home,
+                agent,
+                branch,
+                source_repo_hint.as_deref(),
+            );
+            json!({
+                "released": true,
+                "dir_existed": o.dir_existed,
+                "dir_removed": o.dir_removed,
+                "binding_outcome": o.binding_outcome,
+                "git_metadata_pruned": gc.pruned_count,
+                "git_metadata_repos": gc.repos_touched,
+            })
+        }
+        Err(e) => json!({"error": e, "code": "path_outside_pool"}),
     }
 }
 
 #[cfg(test)]
 #[allow(clippy::unwrap_used)]
-mod path_traversal_tests {
-    use super::*;
-    use serde_json::json;
+mod tests;
 
-    #[test]
-    fn source_repo_with_dotdot_rejected() {
-        let home = std::env::temp_dir().join("pt-test-1");
-        std::fs::create_dir_all(&home).ok();
-        let args = json!({"branch": "feat-x", "repository_path": "/tmp/../etc/passwd"});
-        let sender = Some(crate::identity::Sender::new("agent-1").unwrap());
-        let result = handle_bind_self(&home, &args, &sender);
-        assert_eq!(result["code"].as_str(), Some("path_traversal"));
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn nested_traversal_rejected() {
-        let home = std::env::temp_dir().join("pt-test-2");
-        std::fs::create_dir_all(&home).ok();
-        let args = json!({"branch": "feat-x", "repository_path": "/home/user/foo/../../etc"});
-        let sender = Some(crate::identity::Sender::new("agent-2").unwrap());
-        let result = handle_bind_self(&home, &args, &sender);
-        assert_eq!(result["code"].as_str(), Some("path_traversal"));
-        std::fs::remove_dir_all(&home).ok();
-    }
-}
+#[cfg(test)]
+#[allow(clippy::unwrap_used)]
+mod path_traversal_tests;
